requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.4
djangorestframework==3.16.0
frozenlist==1.7.0
idna==3.10
kombu==5.5.4
multidict==6.6.3
packaging==25.0
prompt_toolkit==3.0.51
propcache==0.3.2
python-dateutil==2.9.0.post0
redis==6.2.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
yarl==1.20.1